refactor: CarRace 생성자 변경
