+++ src/test/java/racingcar/util/NumberGeneratorTest.java
@@ -2,6 +2,8 @@ package racingcar.util;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.RepeatedTest;
+import racingcar.domain.NumberGenerator;
+import racingcar.domain.RandomNumberGenerator;
 
 import static org.assertj.core.api.Assertions.assertThat;
